Used pcall to detect user error and handle it
Wanted a way to catch the user entering an invalid operand. Didn't want
the error to print out a traceback message, just let the user know they
made a mistake and would have to start again. Found wrapping the variable
assignment into a function, I could use pcall to check that function for
errors, and handle them in the way I wanted.