Fixed a few Lua warnings
Didn't need to assign "Loc" to a default
as it gets assigned in the for function.
Still has a warning about it being nil
in line 40, where it still has no value
but it get populated when run, so the
warning can be ignored.

Also made a comment explaining that.