Fixed a warning about integer having nil value
Kept getting a warning where I tried to reuse a variable in a
function to show where something was located. Still not sure
why I was getting the warning, the programs worked despite it,
but I changed the location to "-1" rather than the variable.
That cleared up the warning, and everything still works.